test: extract shared TaskStubTrait to cut duplication
The per-file createTaskStub helpers in the unit/integration/e2e template
tests were near-identical; SonarCloud flagged the new-code duplication.
Extract one shared TaskStubTrait (Tests/Support) and consume it from all
three suites instead of relocating the clone per file.
